% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{get_example_data}
\alias{get_example_data}
\title{"Download the cellphoneDB data in your current directory from this package."}
\usage{
get_example_data()
}
\value{
A file path which contains the example cellphone DB files.
}
\description{
"Download the cellphoneDB data in your current directory from this package."
}
\examples{
{
cat(CytoTalknew::get_example_data())
}
}
