% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lrp_mouse}
\alias{lrp_mouse}
\title{Mouse ligand-receptor pairs}
\format{
A data frame with two variables: \code{ligand} and \code{receptor}
}
\source{
\url{https://www.nature.com/articles/s41586-018-0590-4}
}
\usage{
lrp_mouse
}
\description{
Contains the names of mouse ligand-receptor pairs
}
\note{
DOI: \doi{10.1038/s41586-018-0590-4}.
}
\examples{
{
data(lrp_mouse,package="CytoTalknew") }      # Load the data
}
\keyword{datasets}
