% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcg_human}
\alias{pcg_human}
\title{Human protein-coding genes (PCG)}
\format{
A large character vector
}
\source{
\url{https://www.nature.com/articles/s41586-018-0590-4}
}
\usage{
pcg_human
}
\description{
Contains the names of human protein-coding genes
}
\note{
DOI: \doi{10.1038/s41586-018-0590-4}
}
\examples{
{
data(pcg_human,package="CytoTalknew") }      # Load the data
}
\keyword{datasets}
