% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio.R
\name{read_matrix_with_meta}
\alias{read_matrix_with_meta}
\title{Read scRNAseq Data Matrix and Metadata}
\usage{
read_matrix_with_meta(
  fpath_mat,
  fpath_meta,
  auto_transform = TRUE,
  silent = TRUE
)
}
\arguments{
\item{fpath_mat}{The path of a file containing a scRNAseq data matrix}

\item{fpath_meta}{The path of a file contianing column metadata (cell types)}

\item{auto_transform}{Should count data be transformed if detected?}
}
\value{
A named list containing a sparse data matrix and cell type metadata
}
\description{
Read scRNAseq Data Matrix and Metadata
}
\examples{
{
base_path <- CytoTalknew::get_example_data()
fpath_mat <- file.path(base_path, "sample_counts.txt")
fpath_meta <- file.path(base_path, "sample_meta.txt")
lst_scrna <- CytoTalknew::read_matrix_with_meta(fpath_mat, fpath_meta)
table(lst_scrna$cell_types)
}

}
