% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{result_cyto}
\alias{result_cyto}
\title{Example Pipeline Result (CytoTalk)}
\format{
A list
}
\usage{
result_cyto
}
\description{
Example Pipeline Result (CytoTalk)
}
\examples{
{
data(result_cyto,package="CytoTalknew")   }    # Load the data
}
\keyword{datasets}
