% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrna_cpdb}
\alias{scrna_cpdb}
\title{Example scRNAseq Data (CellphoneDB)}
\format{
A list
}
\source{
\url{https://www.nature.com/articles/s41586-018-0590-4}
}
\usage{
scrna_cpdb
}
\description{
Example scRNAseq Data (CellphoneDB)
}
\note{
DOI: \doi{10.1038/s41586-018-0590-4}
}
\examples{
{
data(scrna_cpdb,package="CytoTalknew") }      # Load the data
}
\keyword{datasets}
