% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrna_cyto}
\alias{scrna_cyto}
\title{Example scRNAseq Data (CytoTalk)}
\format{
A list
}
\source{
\url{https://www.nature.com/articles/s41586-018-0590-4}
}
\usage{
scrna_cyto
}
\description{
Example scRNAseq Data (CytoTalk)
}
\note{
DOI: \doi{10.1038/s41586-018-0590-4}
}
\examples{
{
data(scrna_cyto,package="CytoTalknew")   }    # Load the data
}
\keyword{datasets}
